LayoutTests/imported/w3c/web-platform-tests/fetch/sec-metadata/xslt.tentative.https.sub-expected.txt




FAIL Same-Origin xslt assert_not_equals: Empty Sec-Metadata header. got disallowed value ""
FAIL Same-site xslt assert_equals: destination expected (string) "xslt" but got (undefined) undefined
FAIL Cross-site xslt assert_equals: destination expected (string) "xslt" but got (undefined) undefined
